split up ConfigurePatch into chunks that can handle loading & state logic
