+++ package.json
@@ -49,7 +49,7 @@
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0",
+    "react-scripts": "3.4.1",
     "typescript": "^3.7.4",
     "waait": "^1.0.5"
   },
